execute function to array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,8 @@
 #include "function_pointers.h"
+#include <stddef.h>
 
 /**
- * array_iterator - iterates and executes 
+ * array_iterator - iterates and executes
  * function pointer
  *
  * @array: array
@@ -11,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
